pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
requires-python=">=3.9"
name = "moviepy"
version = "2.2.0"
description = "Video editing with Python"
readme = "README.md"
license = { text = "MIT License" }
authors = [{ name = "Zulko 2025" }]
keywords = ["video", "editing", "audio", "compositing", "ffmpeg"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
    "Topic :: Multimedia :: Video",
    "Topic :: Multimedia :: Video :: Capture",
    "Topic :: Multimedia :: Video :: Conversion",
]
dependencies = [
    "decorator>=4.0.2,<6.0",
    "imageio>=2.5,<3.0",
    "imageio_ffmpeg>=0.2.0",
    "numpy>=1.25.0",
    "proglog<=1.0.0",
    "python-dotenv>=0.10",
    "pillow>=9.2.0",
    "opencv-python-headless>=4.0.0",
]

[project.optional-dependencies]
accelerate = [
    "numba>=0.59",
]

# GPU compositing is optional and best-effort.
# We ship CuPy GPU kernels; OpenCV-CUDA support requires a custom OpenCV build
# (PyPI wheels generally don't include CUDA). We still keep OpenCV as a base
# dependency for CPU video decoding.
gpu = [
    "cupy-cuda13x>=13.0.0",
    # Optional: NVIDIA's pip-distributed Video Codec SDK bindings (encode/decode).
    # Note: runtime requires compatible NVIDIA drivers / CUDA libraries.
    "pynvvideocodec>=2.0.0; platform_system != 'Darwin'",
    # On Windows, PyNvVideoCodec wheels currently depend on the CUDA 12 runtime DLL
    # `cudart64_12.dll`. Installing this package provides it.
    "nvidia-cuda-runtime-cu12>=12.0; platform_system == 'Windows'",
]
doc = [
    "numpydoc<2.0",
    "Sphinx==6.*",
    "pydata-sphinx-theme==0.13",
    "sphinx_design",
    # PyNvVideoCodec wheels depend on CUDA 12 runtime DLLs (e.g. cudart64_12.dll).
    # On Windows, this is most reliably satisfied via the pip runtime package.
    "nvidia-cuda-runtime-cu12>=12.0.0; platform_system == 'Windows'",
]
test = [
    "coveralls>=3.0,<4.0",
    "pytest-cov>=2.5.1,<3.0",
    "pytest>=3.0.0,<7.0.0",
]
lint = [
    "black>=23.7.0",
    "flake8>=6.0.0",
    "flake8-absolute-import>=1.0",
    "flake8-docstrings>=1.7.0",
    "flake8-rst-docstrings>=0.3",
    "flake8-implicit-str-concat==0.4.0",
    "isort>=5.12",
    "pre-commit>=3.3",
]

[tool.setuptools.packages.find]
include = ["moviepy*"]
exclude = ["media", "tests", "docs"]
